+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* 24/09/22
+    - 0.0.2 (utils/email.py) <br>
+        consider mobile <br>
+
 * 24/09/21
     - 0.0.1 <br>
         service start <br>